fixed some runner issues
